backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
facenet-pytorch==2.5.3
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.26.2
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1
